requirements.txt
requests>=2.22.0
numpy>=1.16.5
pandas>=1.0.0
scikit-learn>=0.22
python-dateutil>=2.8.0
tensorflow==2.3.0
pykrige==1.5.1
geopandas==0.8.1
tqdm >= 4.41
